+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+- [FEATURE] Support WebView recording in Session Replay. See [#1776][]
+
 # 2.10.0 / 23-04-2024
 
 - [IMPROVEMENT] Add image duplicate detection between sessions. See [#1747][]
@@ -631,16 +633,17 @@ Release `2.0` introduces breaking changes. Follow the [Migration Guide](MIGRATIO
 [#1697]: https://github.com/DataDog/dd-sdk-ios/pull/1697
 [#1707]: https://github.com/DataDog/dd-sdk-ios/pull/1707
 [#1711]: https://github.com/DataDog/dd-sdk-ios/pull/1711
+[#1721]: https://github.com/DataDog/dd-sdk-ios/pull/1721
 [#1722]: https://github.com/DataDog/dd-sdk-ios/pull/1722
 [#1724]: https://github.com/DataDog/dd-sdk-ios/pull/1724
 [#1741]: https://github.com/DataDog/dd-sdk-ios/pull/1741
 [#1742]: https://github.com/DataDog/dd-sdk-ios/pull/1742
 [#1746]: https://github.com/DataDog/dd-sdk-ios/pull/1746
+[#1747]: https://github.com/DataDog/dd-sdk-ios/pull/1747
 [#1774]: https://github.com/DataDog/dd-sdk-ios/pull/1774
 [#1763]: https://github.com/DataDog/dd-sdk-ios/pull/1763
 [#1767]: https://github.com/DataDog/dd-sdk-ios/pull/1767
-[#1721]: https://github.com/DataDog/dd-sdk-ios/pull/1721
-[#1747]: https://github.com/DataDog/dd-sdk-ios/pull/1747
+[#1776]: https://github.com/DataDog/dd-sdk-ios/pull/1776
 [@00fa9a]: https://github.com/00FA9A
 [@britton-earnin]: https://github.com/Britton-Earnin
 [@hengyu]: https://github.com/Hengyu